Remove hdm and dr tests
[ci]
