+++ .github/workflows/build.yml
@@ -163,20 +163,6 @@ jobs:
       oss_only: ${{ inputs.oss_only }}
     secrets: inherit
 
-  scala_test_disaster_recovery:
-    uses: ./.github/workflows/build.scala_test.yml
-    with:
-      runs_on: self-hosted-k8s-x-large
-      test_names_file: 'test-full-class-names-disaster-recovery.log'
-      start_canton_options: -w
-      parallelism: 1
-      test_name: disaster-recovery
-      commit_sha: ${{ inputs.commit_sha }}
-      daml_base_version: ${{ inputs.daml_base_version }}
-      protocol_version: ${{ inputs.protocol_version }}
-      oss_only: ${{ inputs.oss_only }}
-    secrets: inherit
-
   scala_test_logical_sync_upgrade:
     uses: ./.github/workflows/build.scala_test.yml
     with:
@@ -417,7 +403,6 @@ jobs:
     - scala_test_sim_time
     - scala_test_resource_intensive
     - scala_test_record_time_tolerance
-    - scala_test_disaster_recovery
     - scala_test_logical_sync_upgrade
     - scala_test_roll_forward_lsu
     - scala_test_with_cometbft
