+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/RollForwardLsuIntegrationTest.scala
@@ -70,7 +70,7 @@ class RollForwardLsuIntegrationTest
                 // TODO(#4682) Make it work with BFT connections.
                 bftSequencerConnection = false,
                 domainMigrationDumpPath = Some(
-                  (DomainMigrationUtil.migrationTestDumpDir(
+                  (SynchronizerUpgradeUtil.migrationTestDumpDir(
                     name
                   ) / "domain_migration_dump.json").path
                 ),
