+++ build.sbt
@@ -2117,7 +2117,6 @@ updateTestConfigForParallelRuns := {
   def isDockerComposeValidatorPreflightIntegrationTest(name: String): Boolean =
     isPreflightIntegrationTest(name) && name.contains("DockerComposeValidator")
 
-  def isDisasterRecoveryTest(name: String): Boolean = name contains "DisasterRecovery"
   def isAppUpgradeTest(name: String): Boolean = name contains "AppUpgrade"
   // These are tests that are particularly resource intensive and need larger runners.
   // Usually that is because they need to spin up an additional Canton instance within the test.
@@ -2236,11 +2235,6 @@ updateTestConfigForParallelRuns := {
       "Non-DevNet Preflight tests against runbook validator",
       "test-full-class-names-validator-preflight-non-devnet.log",
       (t: String) => isRunbookValidatorPreflightIntegrationTest(t) && isNonDevNetTest(t),
-    ),
-    (
-      "disaster recovery tests",
-      "test-full-class-names-disaster-recovery.log",
-      (t: String) => !isTimeBasedTest(t) && isDisasterRecoveryTest(t),
     ),
     (
       "app upgrade tests",
