--- test-full-class-names-disaster-recovery.log
@@ -1,1 +1,0 @@
-org.lfdecentralizedtrust.splice.integration.tests.DisasterRecoveryIntegrationTest
